fix(saturn,sh2): silence new clippy lints from the rust 1.95 toolchain bump
A newer clippy newly flags pre-existing code: drop the no-op `0 << 8` /
`+ 0x000` identity ops (cd_block CR identity string, VDP2 TVMD test
writes), the redundant `u32 -> u32` cast in trace_boot, and allow
`single_range_in_vec_init` in rom_harness (a single-region `&[a..b]` slice
into `state_digest` is intentional, not a mis-built range-vec). No
behaviour change; restores `cargo clippy --workspace -- -D warnings`.

+++ crates/saturn/tests/trace_boot.rs
@@ -1119,7 +1119,7 @@ fn catch_divergent_memcmp() {
     println!(
         "memcmp entry at step {steps}: R7=0x{r7:08X} R5=0x{r5:08X} len R6=0x{r6:08X} PR=0x{pr:08X}"
     );
-    let n = (r6.min(24)) as u32;
+    let n = r6.min(24);
     let dump = |sat: &mut Saturn, base: u32, label: &str| {
         let mut s = String::new();
         for i in 0..n {
